refactor: create a dedicated Ark/VM/Value/ folder for better organization
