+++ include/Ark/Builtins/Builtins.hpp
@@ -13,7 +13,7 @@
 
 #include <vector>
 
-#include <Ark/VM/Value.hpp>
+#include <Ark/VM/Value/Value.hpp>
 
 namespace Ark
 {
