+++ include/Ark/TypeChecker.hpp
@@ -17,7 +17,7 @@
 #include <sstream>
 
 #include <Ark/Error/Exceptions.hpp>
-#include <Ark/VM/Value.hpp>
+#include <Ark/VM/Value/Value.hpp>
 
 namespace Ark
 {
