+++ include/Ark/VM/ScopeView.hpp
@@ -1,5 +1,5 @@
 /**
- * @file Scope.hpp
+ * @file ScopeView.hpp
  * @author Lexy Plateau (lexplt.dev@gmail.com)
  * @brief The virtual machine scope system
  * @date 2020-10-27
@@ -15,7 +15,7 @@
 #include <cinttypes>
 
 #include <Ark/Utils/Platform.hpp>
-#include <Ark/VM/Value.hpp>
+#include <Ark/VM/Value/Value.hpp>
 
 namespace Ark::internal
 {
