+++ include/Ark/VM/Value/Future.hpp
@@ -14,7 +14,7 @@
 #include <future>
 #include <vector>
 
-#include <Ark/VM/Value.hpp>
+#include <Ark/VM/Value/Value.hpp>
 #include <Ark/VM/ExecutionContext.hpp>
 
 namespace Ark
